drop torchvision, tests only
