+++ pl_examples/requirements.txt
@@ -1,0 +1,1 @@
+torchvision>=0.4.0
