--- pytorch_lightning/testing/__init__.py
@@ -1,12 +1,0 @@
-from .model import LightningTestModel
-from .model_base import LightningTestModelBase
-from .model_mixins import (
-    LightningValidationStepMixin,
-    LightningValidationMixin,
-    LightningValidationStepMultipleDataloadersMixin,
-    LightningValidationMultipleDataloadersMixin,
-    LightningTestStepMixin,
-    LightningTestMixin,
-    LightningTestStepMultipleDataloadersMixin,
-    LightningTestMultipleDataloadersMixin,
-)
