+++ requirements.txt
@@ -2,7 +2,6 @@ scikit-learn>=0.20.2
 tqdm>=4.35.0
 numpy>=1.16.4
 torch>=1.1
-torchvision>=0.4.0, < 0.5  # the 0.5. has some issues with torch JIT
 tensorboard>=1.14
 future>=0.17.1  # required for builtins in setup.py
