+++ tests/requirements.txt
@@ -1,3 +1,4 @@
+torchvision>=0.4.0, < 0.5  # the 0.5. has some issues with torch JIT
 tox
 coverage
 codecov
