+++ tests/test_trainer.py
@@ -8,7 +8,7 @@ from pytorch_lightning import Trainer
 from pytorch_lightning.callbacks import (
     ModelCheckpoint,
 )
-from pytorch_lightning.testing import (
+from tests.models import (
     LightningTestModel,
     LightningTestModelBase,
     LightningValidationStepMixin,
